+++ src/cryptonote_basic/connection_context.h
@@ -47,7 +47,7 @@ namespace cryptonote
   {
     cryptonote_connection_context(): m_state(state_before_handshake), m_remote_blockchain_height(0), m_last_response_height(0),
         m_expected_heights_start(0), m_last_request_time(boost::date_time::not_a_date_time), m_callback_request_count(0),
-        m_last_known_hash(crypto::null_hash), m_pruning_seed(0), m_rpc_port(0), m_rpc_credits_per_hash(0), m_anchor(false), m_score(0),
+        m_last_known_hash(crypto::null_hash), m_pruning_seed(0), m_anchor(false), m_score(0),
         m_expect_response(0), m_expect_height(0), m_num_requested(0) {}
 
     enum state
@@ -111,8 +111,6 @@ namespace cryptonote
     copyable_atomic m_callback_request_count; //in debug purpose: problem with double callback rise
     crypto::hash m_last_known_hash;
     uint32_t m_pruning_seed;
-    uint16_t m_rpc_port;
-    uint32_t m_rpc_credits_per_hash;
     bool m_anchor;
     int32_t m_score;
     int m_expect_response;