+++ src/p2p/p2p_protocol_defs.h
@@ -73,16 +73,12 @@ namespace nodetool
     peerid_type id;
     int64_t last_seen;
     uint32_t pruning_seed;
-    uint16_t rpc_port;
-    uint32_t rpc_credits_per_hash;
 
     BEGIN_KV_SERIALIZE_MAP()
       KV_SERIALIZE(adr)
       KV_SERIALIZE(id)
       KV_SERIALIZE_OPT(last_seen, (int64_t)0)
       KV_SERIALIZE_OPT(pruning_seed, (uint32_t)0)
-      KV_SERIALIZE_OPT(rpc_port, (uint16_t)0)
-      KV_SERIALIZE_OPT(rpc_credits_per_hash, (uint32_t)0)
     END_KV_SERIALIZE_MAP()
 
     BEGIN_SERIALIZE()
@@ -90,8 +86,6 @@ namespace nodetool
       FIELD(id)
       VARINT_FIELD(last_seen)
       VARINT_FIELD(pruning_seed)
-      VARINT_FIELD(rpc_port)
-      VARINT_FIELD(rpc_credits_per_hash)
     END_SERIALIZE()
   };
   typedef peerlist_entry_base<epee::net_utils::network_address> peerlist_entry;
@@ -147,9 +141,7 @@ namespace nodetool
     for(const peerlist_entry& pe: pl)
     {
       ss << peerid_to_string(pe.id) << "\t" << pe.adr.str()
-        << " \trpc port " << (pe.rpc_port > 0 ? std::to_string(pe.rpc_port) : "-")
-        << " \trpc credits per hash " << (pe.rpc_credits_per_hash > 0 ? std::to_string(pe.rpc_credits_per_hash) : "-")
-        << " \tpruning seed " << pe.pruning_seed 
+        << " \tpruning seed " << pe.pruning_seed
         << " \tlast_seen: " << (pe.last_seen == 0 ? std::string("never") : epee::misc_utils::get_time_interval_string(now_time - pe.last_seen))
         << std::endl;
     }
@@ -179,10 +171,8 @@ namespace nodetool
 
   struct basic_node_data
   {
-    uuid network_id;                   
+    uuid network_id;
     uint32_t my_port;
-    uint16_t rpc_port;
-    uint32_t rpc_credits_per_hash;
     peerid_type peer_id;
     uint32_t support_flags;
 
@@ -190,8 +180,6 @@ namespace nodetool
       KV_SERIALIZE_VAL_POD_AS_BLOB(network_id)
       KV_SERIALIZE(peer_id)
       KV_SERIALIZE(my_port)
-      KV_SERIALIZE_OPT(rpc_port, (uint16_t)(0))
-      KV_SERIALIZE_OPT(rpc_credits_per_hash, (uint32_t)0)
       KV_SERIALIZE_OPT(support_flags, (uint32_t)0)
     END_KV_SERIALIZE_MAP()
   };